requirements.txt
apiai==1.2.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
geojson==2.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
pyowm==2.10.0
python-decouple==3.1
requests==2.21.0
simplejson==3.16.0
urllib3==1.24.1
Werkzeug==0.14.1
